[maybe] Add package prefix to the eponymous macros as well
NOTE Only do this if a cond-let* is actually added to Emacs.

In the initial release of this package `cond-let' and `cond-let*'
were provided exactly under these names (without the `cond-let--'
prefix), because these forms did not exist in Emacs at that time.
Later an incompatible implementation of `cond-let*' was added to
Emacs, forcing us to rename our implementation.  To avoid having
a single un-prefixed form, we also add the prefix to `cond-let'.